Add TASK_DYNAMICS_DERIVED_FACTORY_Z()
This is a special case for TASK_DYNAMICS_DERIVED_FACTORY(...) when no
arguments are provided. This avoid relying on non standard extensions in
a public header -- which breaks compilation if TVM is used with a
stricter standard, e.g. -std=c++17 instead of -std=gnu++17 on Ubuntu 20.04

A future (nicer) solution would be to use __VA_OPT__ but this is a C++20
feature.

However, it's relatively unlikely this macro will see much use as task
dynamics without construction parameters can be expected to be uncommon.