Generalize GraphProbe::registerAccessor
- Only one function
- Support noexcept out of the box

The cost is that it might require to static_cast the accessor in case of
overload (the compiler cannot decide what is meant, even in presence of
SFINAE). In practice it is expected to be extremely rare because TVM
output are rarely settable.

Static cast are required in
registerTVMFunction()/registerTVMConstraint()/registerTVMTaskDynamics()
in case of overload issues (which only happens for
BasicLinearConstraint::l and BasicLinearConstraint::u inside TVM itself)